fix bbox list int

+++ src/lib/datasets/core/heatmap_target.py
@@ -91,7 +91,7 @@ def get_set_hoi_target(hoi_list, bbox_list, output_w, output_h, max_rels, num_cl
     hm_rel = torch.zeros(max_rels, dtype=torch.long)
     hoi_list = torch.tensor(hoi_list)[:num_rels, :]
     sub_obj_bbox_id, hm_rel[:num_rels] = hoi_list[:, :2], hoi_list[:, -1]
-    bbox_list = torch.tensor(bbox_list)
+    bbox_list = torch.tensor(bbox_list).int()
     matrix = torch.stack([torch.arange(0, output_w).repeat(output_h, 1),
                           torch.arange(0, output_h).repeat(output_w, 1).T], dim=0).repeat(max_rels, 2, 1, 1)
     sub_obj_ct = torch.zeros(max_rels, 4, 1, 1)
